+++ src/envoy/http/mixer/filter.h
@@ -19,6 +19,7 @@
 #include "envoy/access_log/access_log.h"
 #include "envoy/http/filter.h"
 #include "src/envoy/http/mixer/control.h"
+#include "src/envoy/utils/message_counter.h"
 
 namespace Envoy {
 namespace Http {
@@ -31,30 +32,6 @@ struct PerRouteServiceConfig : public Router::RouteSpecificFilterConfig {
 
   // Its config hash
   std::string hash;
-};
-
-// The struct to store gRPC message counter state.
-struct GrpcMessageCounter {
-  GrpcMessageCounter() : state(ExpectByte0), current_size(0), count(0){};
-
-  // gRPC uses 5 byte header to encode subsequent message length
-  enum GrpcReadState {
-    ExpectByte0 = 0,
-    ExpectByte1,
-    ExpectByte2,
-    ExpectByte3,
-    ExpectByte4,
-    ExpectMessage
-  };
-
-  // current read state
-  GrpcReadState state;
-
-  // current message size
-  uint64_t current_size;
-
-  // message counter
-  uint64_t count;
 };
 
 class Filter : public StreamFilter,
@@ -126,8 +103,8 @@ class Filter : public StreamFilter,
 
   // True for gRPC requests
   bool grpc_request_{false};
-  GrpcMessageCounter grpc_request_counter_;
-  GrpcMessageCounter grpc_response_counter_;
+  Utils::GrpcMessageCounter grpc_request_counter_;
+  Utils::GrpcMessageCounter grpc_response_counter_;
 
   // The stream decoder filter callback.
   StreamDecoderFilterCallbacks* decoder_callbacks_{nullptr};
