+++ src/istio/utils/attribute_names.cc
@@ -48,9 +48,9 @@ const char AttributeName::kRequestUserAgent[] = "request.useragent";
 const char AttributeName::kRequestApiKey[] = "request.api_key";
 
 const char AttributeName::kRequestGrpcMessageCount[] =
-    "request.grpc_message_count";
+    "request.grpc.message_count";
 const char AttributeName::kResponseGrpcMessageCount[] =
-    "response.grpc_message_count";
+    "response.grpc.message_count";
 
 const char AttributeName::kResponseCode[] = "response.code";
 const char AttributeName::kResponseDuration[] = "response.duration";
